Initial Version of the Shelly Reader

+++ src/main/java/edu/hm/greenit/tools/shelly/Meter.java
@@ -18,7 +18,7 @@ public class Meter {
     private double overpower;
     private boolean is_valid;
     private long timestamp;
-    private long total;
+    private double total;
 
     public double getPower() {
         return power;
@@ -44,11 +44,11 @@ public class Meter {
         this.timestamp = timestamp;
     }
 
-    public long getTotal() {
+    public double getTotal() {
         return total;
     }
 
-    public void setTotal(long total) {
+    public void setTotal(double total) {
         this.total = total;
     }
